# Conflicts:
#	engine/utils.py

+++ engine/utils.py
@@ -1,10 +1,10 @@
 import getpass
-
+import sys
 from crontab import CronTab
-
 from engine.models import AllRdsInstanceTypes, AllEc2InstanceTypes, RDS
 from django.db.models import F
 import json
+from django.conf import settings
 
 
 def get_instance_types(cluster_type):
@@ -23,9 +23,13 @@ def get_selection_list(query, table_col, value_col, data_col):
 
 
 def create_cron(rule):
+    if sys.platform == "win32":
+        return
+
     cron = CronTab(user=getpass.getuser())
     cron.remove_all(comment="rule_{}".format(rule.id))
-    job = cron.new(command="python manage.py apply_rule {}".format(rule.id), comment="rule_{}".format(rule.id))
+    job = cron.new(command="{0}/venv/bin/python {0}/manage.py apply_rule {1}".format(
+        settings.BASE_DIR, rule.id), comment="rule_{}".format(rule.id))
 
     # Run at
     job.setall(rule.run_at)
@@ -38,4 +42,14 @@ def create_cron(rule):
     #     job.hour.on(hour)
     # if minute:
     #     job.minute.on(minute)
+    cron.write()
+
+
+def delete_cron(rule):
+    if sys.platform == "win32":
+        return
+
+    cron = CronTab(user=getpass.getuser())
+    cron.remove_all(comment="rule_{}".format(rule.id))
+
     cron.write()
